refactor(github): split enterprise-specific code
